flash_attn

+++ src/kernels.cu
@@ -245,7 +245,7 @@ void flashAttention(const std::vector<T>& h_q, const std::vector<T>& h_k,
     dim3 block(512);
     dim3 grid(CEIL(target_seq_len, Br), query_heads, batch_size);
     size_t sram_size = (Br + Bc) * head_dim * 2 + Br;
-    size_t sram_bytes = sram_size * sizeof(T) + Br * sizeof(float);
+    int sram_bytes = sram_size * sizeof(T) + Br * sizeof(float);
     sram_bytes = min(sram_bytes, max_sram_bytes);
     flash_attn_kernel<T, Br, Bc><<<grid, block, sram_bytes>>>(d_q, d_k, d_v, d_o, target_seq_len, src_seq_len, kv_heads, head_dim, is_causal, 1.0 / sqrt(head_dim));
     cudaMemcpy(h_o.data(), d_o, qo_bytes, cudaMemcpyDeviceToHost);
